smart_city_joinville/requirements.txt
neo4j==5.24.0
langchain==0.3.1
langchain-community==0.3.1
mistralai==1.1.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
flask==3.0.3